+++ main.go
@@ -1,22 +1,17 @@
 package main
 
 import (
-	"database/sql"
 	"fmt"
+	"go-portfolio/db"
+	"go-portfolio/router"
 	"log"
 	"net/http"
+	"os"
 	"time"
 
-	"github.com/gorilla/mux"
 	"github.com/joho/godotenv"
 	_ "github.com/lib/pq"
 )
-
-// App holds the router and the db instances
-type App struct {
-	db     *sql.DB
-	router *mux.Router
-}
 
 func main() {
 	envErr := godotenv.Load()
@@ -25,18 +20,23 @@ func main() {
 		log.Fatal(envErr)
 	}
 
-	a := App{}
 	address := ":8080"
 
-	a.db = NewDataBase()
-	defer a.db.Close()
+	database := db.New(
+		os.Getenv("DB_HOST"),
+		os.Getenv("DB_PORT"),
+		os.Getenv("DB_USER"),
+		os.Getenv("DB_PASSWORD"),
+		os.Getenv("DB_NAME"),
+	)
+	defer database.Close()
 
-	a.SetUpRouter()
+	r := router.New(database)
 
 	fmt.Println("Running App on ", address)
 
 	server := &http.Server{
-		Handler:      a.router,
+		Handler:      r,
 		Addr:         address,
 		WriteTimeout: 15 * time.Second,
 		ReadTimeout:  15 * time.Second,
